Move PropertyHandler and HashProperties
