+++ src/base/url/HashProperties.ts
